Modified lseek block-handling code

+++ current/vmlarix/filesystem/sfs/sfs_lseek.c
@@ -38,7 +38,7 @@ int sfs_lseek(filedesc *f, off_t offset, int whence)
   uint32_t blksize = fstat_buf->st_blksize;
   uint32_t fsize = fstat_buf->st_size;
   kfree(fstat_buf);
-  uint32_t newpos;
+  uint64_t newpos;
   if(whence==SEEK_SET)
     newpos = offset;
   else if(whence==SEEK_CUR)
